rename(LineStrategy): 라인을 생성하는 Strategy 클래스를 util 패키지에서 domain 패키지 내로 이동
