+++ src/main/java/ladder/domain/strategy/linestrategy/LineStrategy.java
@@ -1,4 +1,4 @@
-package ladder.util;
+package ladder.domain.strategy.linestrategy;
 
 import ladder.domain.Line;
